# Conflicts:
#	app/src/main/java/com/drexelsp/blunote/blunote/Player.java
#	app/src/main/java/com/drexelsp/blunote/ui/LoginActivity.java

+++ app/src/main/java/com/drexelsp/blunote/blunote/Player.java
@@ -52,6 +52,11 @@ public class Player extends Observable implements Runnable, MediaPlayer.OnComple
         EventBus.getDefault().register(this);
     }
 
+    /**
+     * Synchronized method to add a Blunote.Song to the player's queue to be played.
+     *
+     * @param song
+     */
     public AtomicBoolean getIsPaused() {
         return isPaused;
     }
@@ -74,7 +79,7 @@ public class Player extends Observable implements Runnable, MediaPlayer.OnComple
      * @return millis
      */
     public int getCurrentMillisecond() {
-        return player.getCurrentPosition();
+        return (int) player.getCurrentPosition();
     }
